requirements.txt
requests~=2.28.2
pika~=1.3.1
pytest~=7.2.2
pytest-html-reporter
p
docker~=6.0.1
pyodbc~=4.0.35
python-dotenv~=1.0.0